+++ changelog.md
@@ -96,6 +96,8 @@ WIP: big text
 # v1.4
 + restructured code to be more flexible (structure was hindering the management of multi-continuous chat messages and multi-link messages)
 + changed formatting for _be command to include 'k/d'
++ added formatting to 'Your role does not have requirements!'
+